refactor : jwt 관련 리팩토링 완료.
